Move enemy engine glow drawing to effects module
